feat: Launch via Steam on Non Windows Systems if possible
